Test of events getting publishes PASSES.
The say command on OSX didn't run when called inside a Command.
Also, /usr/bin/wc -l was seeing zero lines, even though the file
was definitely there.

So, punt on that approach and just use channels (my first use!).
There were a few wrinkles to work out; for example, how do you
know when the channel is drained for the last time?  Answer: the
event sender has to close it and the receiver has to check the
error half of the returned tuple.