Prep for FileSystem store getting event types from registry.
Currenly, we pass list to FileSystem store constructor, but there
is no need for this code; all the events are already registered
so we should just use that registry.

But ... this requires that we ensure all events listeners are
registered before we register the event store.  Add a panic if
this condition is violated, and update the unit tests so they
conform.