Assume every event store needs a list of event types to do its thing.
The event store uses gob, so it needs to register the different
events to that it can reconsitute the concrete types into a list
of interfaces.

I imagine JSON or SQL would need the same mapping; those would
probably map the type name to the object type.

In any case, when you register the event store, it now calls the
new method on the EventStore interface SetEventTypes().