SqliteEventStore creates table for event type.
Uses SetEventTypes hook to inspect structure of the Event implementations
and generates CREATE TABLE sql.  Generates INSERT INTO and SELECT FROM
sql at the same time, and caches it using the TypeOf(event) as the key.

The logic checks if the table already exists, and verifies that if it does
exist it has the same schema.  If not, it panics.