Change example_test to use Aggregator methods with pointer reciver.
The job of an Aggregator is to maintain state so it can enforce
business rules.  Using pointer receivers is the most natural setup,
as Aggregator methods will need to mutate Aggregator state.

Remove assert that Aggregator can't be a pointer.  The New() method
on the Aggregator interface made that check obsolete.